move searchicon to header, some styling, add provider to header-test
