+++ src/__tests__/components/header.test.tsx
@@ -1,8 +1,10 @@
 import React from 'react';
 import { Router } from 'react-router-dom';
 import renderer from 'react-test-renderer';
-import Header from '../../components/header';
+import Header from '../../components/Header/header';
 import { createMemoryHistory } from 'history';
+import store from '../../store/store';
+import { Provider } from 'react-redux';
 const history = createMemoryHistory();
 
 describe('header', () => {
@@ -10,9 +12,11 @@ describe('header', () => {
         it('should render correctly', () => {
             const tree = renderer
                 .create(
-                    <Router history={history}>
-                        <Header />
-                    </Router>,
+                    <Provider store={store}>
+                        <Router history={history}>
+                            <Header />
+                        </Router>
+                    </Provider>,
                 )
                 .toJSON();
             expect(tree).toMatchSnapshot();
