+++ src/pages/groupPage.tsx
@@ -52,12 +52,7 @@ const GroupPage = (): JSX.Element => {
 
     return (
         <div className="container body">
-            <div className="row">
-                <div className="searchIconContainer">
-                    <SearchIcon />
-                </div>
-            </div>
-            <SearchBar/>
+            <SearchBar />
             {reduxState.thunk.fetch_in_progress ? (
                 <div className="center_container">
                     <Spinner animation="border" />
