+++ src/styles/header.css
@@ -3,6 +3,8 @@
     background-color: #ffab40;
     width: 100vw !important;
     max-width: 100% !important;
+    display: flex;
+    justify-content: space-between;
 }
 .headerLink {
     width: fit-content;
